refactor(security): Move transfer file validation from FileSystem to ConnectionManager
